Fix: Rename namespace and class
