+++ src/License/CopyrightYears.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/php-cs-fixer-config
  */
 
-namespace Ergebnis\PhpCsFixer\Config\Header;
+namespace Ergebnis\PhpCsFixer\Config\License;
 
 /**
  * @internal
