+++ src/License/Url.php
@@ -11,12 +11,12 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/php-cs-fixer-config
  */
 
-namespace Ergebnis\PhpCsFixer\Config\Header;
+namespace Ergebnis\PhpCsFixer\Config\License;
 
 /**
  * @internal
  */
-final class License
+final class Url
 {
     private $value;
 
@@ -25,14 +25,25 @@ final class License
         $this->value = $value;
     }
 
+    /**
+     * @param string $value
+     *
+     * @throws \InvalidArgumentException
+     *
+     * @return self
+     */
     public static function fromString(string $value): self
     {
-        return new self(\trim($value));
-    }
+        $trimmed = \trim($value);
 
-    public function isEmpty(): bool
-    {
-        return '' === $this->value;
+        if (false === \filter_var($trimmed, \FILTER_VALIDATE_URL)) {
+            throw new \InvalidArgumentException(\sprintf(
+                'Value "%s" does not appear to be a valid URL.',
+                $value
+            ));
+        }
+
+        return new self($trimmed);
     }
 
     public function toString(): string
