+++ test/Unit/License/UrlTest.php
@@ -11,16 +11,16 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/php-cs-fixer-config
  */
 
-namespace Ergebnis\PhpCsFixer\Config\Test\Unit\Header;
+namespace Ergebnis\PhpCsFixer\Config\Test\Unit\License;
 
-use Ergebnis\PhpCsFixer\Config\Header\Url;
+use Ergebnis\PhpCsFixer\Config\License\Url;
 use Ergebnis\Test\Util\Helper;
 use PHPUnit\Framework;
 
 /**
  * @internal
  *
- * @covers \Ergebnis\PhpCsFixer\Config\Header\Url
+ * @covers \Ergebnis\PhpCsFixer\Config\License\Url
  */
 final class UrlTest extends Framework\TestCase
 {
